Adding sorting to reporting tables
