+++ src/app/modules/reporting/components/patient-details-dialog/patient-details-dialog.component.html
@@ -7,36 +7,36 @@ cdkDragRootElement=".cdk-overlay-pane">
 
         <legend class="legend">{{data.areaName}}</legend>
 
-        <table mat-table #table [dataSource]="patientRecords" class="table">
+        <table mat-table [dataSource]="patientRecords" matSort class="mat-elevation-z8 table">
 
             <ng-container matColumnDef="emergencynumber">
-              <th mat-header-cell *matHeaderCellDef> Emergency number </th>
+              <th mat-header-cell *matHeaderCellDef mat-sort-header> Emergency number </th>
               <td mat-cell *matCellDef="let element">{{element.emergencynumber}}</td>
             </ng-container>
 
             <ng-container matColumnDef="tagnumber">
-              <th mat-header-cell *matHeaderCellDef> Tag number </th>
+              <th mat-header-cell *matHeaderCellDef mat-sort-header> Tag number </th>
               <td mat-cell *matCellDef="let element"> {{element.tagnumber}} </td>
             </ng-container>
 
             <ng-container matColumnDef="species">
-              <th mat-header-cell *matHeaderCellDef> Species </th>
+              <th mat-header-cell *matHeaderCellDef mat-sort-header> Species </th>
               <td mat-cell *matCellDef="let element"> {{element.species}} </td>
             </ng-container>
 
             <ng-container matColumnDef="callername">
-              <th mat-header-cell *matHeaderCellDef> Caller name </th>
+              <th mat-header-cell *matHeaderCellDef mat-sort-header> Caller name </th>
               <td mat-cell *matCellDef="let element"> {{element.callername}} </td>
             </ng-container>
 
 
             <ng-container matColumnDef="number">
-                <th mat-header-cell *matHeaderCellDef> Number </th>
+                <th mat-header-cell *matHeaderCellDef mat-sort-header> Number </th>
                 <td mat-cell *matCellDef="let element"> {{element.number}} </td>
             </ng-container>
 
             <ng-container matColumnDef="calldate">
-                <th mat-header-cell *matHeaderCellDef> Call date </th>
+                <th mat-header-cell *matHeaderCellDef mat-sort-header> Call date </th>
                 <td mat-cell *matCellDef="let element"> {{element.calldate}} </td>
             </ng-container>
