--- src/app/modules/reporting/components/reporting-page.component.html
@@ -1,38 +1,0 @@
-<fieldset class="form-sub-group">
-    <legend class="legend">AAU Reporting</legend>
-
-    <div fxLayout="column" fxLayoutGap="10px" fxLayout.lt-md="row" fxLayoutAlign = "center center">
-
-        <fieldset class="form-sub-group">
-            <div *ngFor = "let data of patientCountData">
-                <p class="legend">Total animal count: {{data.totalPatientCount}}</p>
-            </div>
-        </fieldset>
-
-        <fieldset class="form-sub-group">
-            <p class="legend">Animal count by area</p>
-            <div *ngFor = "let data of patientCountData">
-                <div *ngFor = "let value of data.patientCountInArea">
-                    <p class="link">{{value.area}}: {{value.count}}</p>
-                </div>
-            </div>
-
-        </fieldset>
-
-        <fieldset class="form-sub-group" [formGroup]="surgeryDetails">
-            <p class="legend">Surgery count by date</p>
-
-            <mat-form-field >
-                <mat-label>Select date</mat-label>
-                <input matInput type="date" formControlName="surgeryDate" name = "SurgeryDate"
-                [errorStateMatcher]="errorMatcher" required>
-            </mat-form-field>
-
-                <p class="link" (click)="openSurgeryDetailsDialog()">{{surgeryCount | async}}</p>
-
-        </fieldset>
-
-
-    </div>
-
-</fieldset>
