--- src/app/modules/reporting/components/reporting-page.component.scss
@@ -1,8 +1,0 @@
-.link {
-    cursor: pointer;
-    color: #607d8b;
-    text-decoration: underline;
-    text-align: center;
-    font-weight: bold;
-    font-size: xx-large;
-  }
