+++ src/app/modules/reporting/components/reporting-record/reporting-record.component.html
@@ -21,9 +21,5 @@
                 </div>
         </fieldset>
     </div>
-
-
-
-
-
+    
 </fieldset>
