+++ src/app/modules/reporting/pages/reporting-page/reporting-page.component.html
@@ -1,1 +1,43 @@
-<reporting-record></reporting-record>
+<fieldset class="form-sub-group">
+    <legend class="legend">AAU Reporting</legend>
+
+    <div fxLayout="row" fxLayoutGap="10px" fxLayout.lt-md="row" fxLayoutAlign = "center center">
+        <fieldset class="form-sub-group">
+            <div *ngFor = "let data of patientCountData">
+                <div *ngIf="data.area === 'Total'">
+                    <p class="legend">Total count : {{data.count}}</p>
+                </div>
+            </div>
+        </fieldset>
+    </div>
+
+    <div fxLayout="column" fxLayoutGap="10px" fxLayout.lt-md="row" fxLayoutAlign = "center center">
+        <fieldset class="form-sub-group">
+            <p class="legend">Animal count in each area</p>
+                <div *ngFor = "let value of patientCountData">
+                    <div *ngIf="value.area != 'Total'">
+                    <p (click)="getPatientDetailsByArea(value.area)" class="link">{{value.area}} : {{value.count}}</p>
+                    </div>
+                </div>
+        </fieldset>
+    </div>
+
+    <div fxLayout="column" fxLayoutGap="10px" fxLayout.lt-md="row" fxLayoutAlign = "center center">
+
+        <fieldset class="form-sub-group" [formGroup]="surgeryDetails">
+            <p class="legend">Surgery count by date</p>
+
+            <mat-form-field >
+                <mat-label>Select date</mat-label>
+                <input matInput type="date" formControlName="surgeryDate" name = "SurgeryDate"
+                [errorStateMatcher]="errorMatcher" required>
+            </mat-form-field>
+
+                <p class="link link-large" (click)="openSurgeryDetailsDialog()">{{surgeryCount | async}}</p>
+
+        </fieldset>
+
+
+    </div>
+
+</fieldset>
