+++ src/app/modules/reporting/pages/reporting-page/reporting-page.component.scss
@@ -1,0 +1,11 @@
+.link {
+    cursor: pointer;
+    color: #607d8b;
+    text-decoration: underline;
+    text-align: center;
+    font-weight: bold;
+  }
+
+.link-large{
+  font-size: xx-large;
+}
