+++ src/app/modules/reporting/reporting-page.module.ts
@@ -2,17 +2,21 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MaterialModule } from 'src/app/material-module';
 import { ReportingPageRoutingModule } from './reporting-page-routing.module';
-import { ReportingPageComponent } from './components/reporting-page.component';
 import { SurgeriesByDateDialogComponent } from './components/surgeries-by-date-dialog/surgeries-by-date-dialog.component';
+import { ReportingPageComponent } from './pages/reporting-page/reporting-page.component';
+import { PatientDetailsDialogComponent } from './components/patient-details-dialog/patient-details-dialog.component';
 
 @NgModule({
     declarations: [
         ReportingPageComponent,
-        SurgeriesByDateDialogComponent
+        SurgeriesByDateDialogComponent,
+        PatientDetailsDialogComponent
     ],
     imports: [
         CommonModule,
         ReportingPageRoutingModule,
-        MaterialModule],
+        MaterialModule
+    ],
 })
+
 export class ReportingPageModule {}
